extract ConditionalCompletableFuture from BatchRpcRequest
